requirements.txt
#base
django>=4
djangorestframework>=3

# ovinc
ovinc-client==0.1.2
bk-crypto-python-sdk==1.1.1

# Celery
celery==5.2.2

# Server
uwsgi==2.0.22

# Datetime
arrow==1.2.3

# RSA
pycryptodome==3.15.0

# Profile
pyinstrument==4.4.0
